Working on mini goals now
I set the reason section up. Now the mini goals didn't take me as long to save and load the values, but the loop is applying the earlier values to the later elements.

Next time:
Update reason section so that if someone wants to delete and clear a section blank they can.
Figure out mini goals section, find way to clear the key value at end of loop instead of carrying it throughout entire process