Reason input removal
THIS WAS TOUGH. Took about 2-3 weeks to figure out how to move the list of reasons down the line based on what gets deleted. Still need to figure out how to save checks but that might not be possible for now. NEXT TIME: Align the "WHATDESIGN" key with the reasoninput key