Last update until final changes for portfolio
Stumped on the submenu section storage removal. I've figured out how to store the info, but removing from storage throws off the order of the items after the target. When I get back to this, consider adding finite amount of list items that can be created, this way (as ugly and inefficient as it may be) I can detail every step needed when a specific item is deleted: e.g. if item 2 deleted, move item 3 to 2 move item 4 to 3 etc. and iterate the steps need for each item removal